switch from int to size_t

+++ ChaosGame.cpp
@@ -121,7 +121,7 @@ int main()
         window.draw(text);
         // User instructions ends.
 
-        for (int i = 0; i < vertices.size(); i++)
+        for (size_t i = 0; i < vertices.size(); i++)
         {
             RectangleShape rect(Vector2f(10, 10));
             rect.setPosition(Vector2f(vertices[i].x, vertices[i].y));
